pyproject.toml
[project]
name = "ai-image-captioner"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "einops>=0.8.1",
    "huggingface-hub[hf-xet]>=0.33.0",
    "langchain>=0.3.25",
    "langchain-mistralai>=0.2.10",
    "langgraph>=0.4.8",
    "pillow>=11.2.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.0",
    "timm>=1.0.15",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "transformers>=4.52.4",
]
